+++ example/src/main/java/com/stripe/example/module/StripeIntentViewModel.kt
@@ -45,7 +45,8 @@ internal class StripeIntentViewModel(
 
     fun createSetupIntent(
         country: String,
-        customerId: String? = null
+        customerId: String? = null,
+        supportedPaymentMethods: String? = null
     ) = makeBackendRequest(
         R.string.creating_setup_intent,
         R.string.setup_intent_status
@@ -55,6 +56,10 @@ internal class StripeIntentViewModel(
                 .plus(
                     customerId?.let {
                         mapOf("customer_id" to it)
+                    }.orEmpty()
+                ).plus(
+                    supportedPaymentMethods?.let {
+                        mapOf("supported_payment_methods" to it)
                     }.orEmpty()
                 )
                 .toMutableMap()