+++ payments-core/src/main/java/com/stripe/android/model/MicrodepositType.kt
@@ -1,0 +1,9 @@
+package com.stripe.android.model
+
+enum class MicrodepositType(val value: String) {
+    // Two non-unique micro-deposits to the customer's bank account
+    AMOUNTS("amounts"),
+    // A single micro-deposit sent to the customer's bank account with a unique descriptor code
+    DESCRIPTOR_CODE("descriptor_code"),
+    UNKNOWN("unknown")
+}