+++ payments-core/src/main/java/com/stripe/android/model/PaymentIntent.kt
@@ -143,6 +143,7 @@ data class PaymentIntent internal constructor(
             is StripeIntent.NextActionData.SdkData -> StripeIntent.NextActionType.UseStripeSdk
             is StripeIntent.NextActionData.RedirectToUrl -> StripeIntent.NextActionType.RedirectToUrl
             is StripeIntent.NextActionData.DisplayOxxoDetails -> StripeIntent.NextActionType.DisplayOxxoDetails
+            is StripeIntent.NextActionData.VerifyWithMicrodeposits -> StripeIntent.NextActionType.VerifyWithMicroDeposits
             else -> null
         }