+++ CHANGELOG.md
@@ -8,6 +8,9 @@ This release enables a new configuration object to be defined for StripeCardScan
 ### CardScan
 * [ADDED] [4689](https://github.com/stripe/stripe-android/pull/4689) The `CardImageVerificationSheet` initializer can now take an additional `Configuration` object.
 
+### Other changes
+* [Added] `us_bank_account` PaymentMethod is now supported.
+
 ## 19.2.2 - 2022-03-01
 * [FIXED] [4606](https://github.com/stripe/stripe-android/pull/4606) Keep status bar color in PaymentLauncher